+++ doc/whats_new.rst
@@ -15,6 +15,10 @@ What's new?
 Changelog
 ~~~~~~~~~
 
+- ``get_rejection_threshold`` now also accepts ECoG and SEEG data by `Nikolai Chapochnikov`_ in :github:`#281`
+
+- RANSAC implementation was refactored, works now with `n_jobs>1` and produces consistent results across different number of jobs. Testing on simulated data added. by `Simon Kern`_ in :github:`#280`
+
 - ``autoreject`` now requires ``mne >= 1.0``, by `Mainak Jas`_, `Alex Gramfort`_, and `Stefan Appelhoff`_ in :github:`#267` and :github:`#268`
 
 - Add ``reject_log`` option to :meth:`autoreject.AutoReject.transform` to enable
@@ -23,8 +27,6 @@ Changelog
 
 - Add :meth:`autoreject.RejectLog.save` and :func:`autoreject.read_reject_log` to
   save and load reject logs, by `Alex Rockhill`_ in :github:`#270`
-
-- ``get_rejection_threshold`` now also accepts ECoG and SEEG data by `Nikolai Chapochnikov`_ in :github:`#281`
 
 .. _0.3:
 
@@ -103,3 +105,4 @@ Changelog
 .. _Stefan Appelhoff: https://stefanappelhoff.com/
 .. _Alex Rockhill: https://github.com/alexrockhill
 .. _Nikolai Chapochnikov: https://github.com/chapochn
+.. _Simon Kern: https://github.com/skjerns